last task error

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,9 +6,9 @@
  */
 int main(void)
 {
-	unsigned int fn_2 = 0;
-	unsigned int fn_1 = 1;
-	unsigned int fn;
+	unsigned long long fn_2 = 0;
+	unsigned long long fn_1 = 1;
+	unsigned long long fn;
 	int i = 0;
 	int j = 1;
